Modified http post for new database schema
I can't test the app locally, and I'm not used to dart, so I'd appreciate if someone could test the changes. I also changed it to send the data to the new server, so again, needs testing. I also used track.uri and I'm not sure if that's the right option.