Added tutorial page
It has explanations on the buttons and the valence/arousal grid. It also uses popup dialogs, which I've added (copied from https://api.flutter.dev/flutter/widgets/PopupRoute-class.html). I added the tutorial page to the the main.dart tab controller. I was only able to test it on Edge, so the page might not fit on an Android screen, and I don't know if the user will be able to automatically scroll down, so that needs to be checked.